Update to Mbed OS 6.16.0

+++ TARGET_LPC11U6X/gpio_irq_api.c
@@ -59,7 +59,7 @@ void gpio_irq5(void) {handle_interrupt_in(5);}
 void gpio_irq6(void) {handle_interrupt_in(6);}
 void gpio_irq7(void) {handle_interrupt_in(7);}
 
-int gpio_irq_init(gpio_irq_t *obj, PinName pin, gpio_irq_handler handler, uint32_t id) {
+int gpio_irq_init(gpio_irq_t *obj, PinName pin, gpio_irq_handler handler, uintptr_t id) {
     // PINT only supprt PIO0_*, PIO1_* and from PIO2_0 to PIO2_7 interrupt
     if (pin >= P2_8) return -1;
